requirements.txt
asgiref==3.7.2
crispy-bootstrap5==0.7
Django==4.2.2
django-crispy-forms==2.0
Pillow==9.5.0
python-dotenv==1.0.0
sqlparse==0.4.4
tzdata==2023.3
